requirements/requirements_win.txt
Django==2.2.2
pycryptodome==3.9.8
six>=1.8.0
requests>=2.4.3
xmltodict>=0.11.0
optionaldict>=0.1.0
python-dateutil>=2.5.2
pyzbar>=0.1.8
Pillow>=6.1.0
PyYaml>=5.1.2
django-jsonfield==1.4.1
pika==1.1.0